+++ src/super/i-block/modules/activation/CHANGELOG.md
@@ -9,6 +9,12 @@ Changelog
 > - :house:      [Internal]
 > - :nail_care:  [Polish]
 
+## v3.0.0-rc.98 (2020-11-16)
+
+#### :rocket: New Feature
+
+* Added `beforeDeactivate`, `deactivated`, `beforeActivate`, `activated` events
+
 ## v3.0.0-rc.46 (2020-07-31)
 
 #### :house: Internal